refactor(agents): consolidate AgentWithRelations import from agent.service for improved code organization
