+++ web-app/src/components/agents/agents.tsx
@@ -1,6 +1,6 @@
 import { useTranslations } from "next-intl";
 
-import { AgentWithRelations } from "@/lib/db/extension/agent";
+import { AgentWithRelations } from "@/lib/db/services/agent.service";
 import { AgentListWithAgent } from "@/lib/db/services/agentList.service";
 import { cn } from "@/lib/utils";
